collect the shared functions
