+++ make_dataset.py
@@ -18,6 +18,7 @@ from tqdm import tqdm
 from src.config import *
 from src.load import DlgDwhLoader
 from src.preprocess import clean_msg
+from src.utils import *
 
 
 # ============================
@@ -67,14 +68,6 @@ def to_dataset(posts: list=None) -> list:
         else:
             dataset.append({"tag": tag, "text": corpus})
     return dataset
-
-
-def mkdir_if_not_exist(path: str):
-    """Make directory if not exist"""
-    p = Path(path)
-    if p.exists():
-        return
-    p.mkdir(parents=True, exist_ok=True)
 
 
 def main():
